feat: Complete PrimerCore modularisation
