+++ Modules/PrimerCore/Sources/PrimerCore/Analytics/Models/AppLifecycleEventProperties.swift
@@ -1,0 +1,27 @@
+//
+//  AppLifecycleEventProperties.swift
+//
+//  Copyright © 2026 Primer API Ltd. All rights reserved. 
+//  Licensed under the MIT License. See LICENSE file in the project root for full license information.
+
+import Foundation
+@_spi(PrimerInternal) import PrimerFoundation
+
+@_spi(PrimerInternal) public struct AppLifecycleEventProperties: AnalyticsEventProperties {
+
+    public enum LifecycleType: String, Codable {
+        case backgrounded = "APP_DID_ENTER_BACKGROUND"
+        case foregrounded = "APP_WILL_ENTER_FOREGROUND"
+    }
+
+    public let lifecycleType: LifecycleType
+    public var params: [String: AnyCodable]?
+
+    public init(lifecycleType: LifecycleType) {
+        self.lifecycleType = lifecycleType
+        let sdkProperties = SDKProperties()
+        let dict = try? sdkProperties.asDictionary()
+        let data = try? JSONSerialization.data(withJSONObject: dict as Any, options: .fragmentsAllowed)
+        data.map { params = try? JSONDecoder().decode([String: AnyCodable].self, from: $0) }
+    }
+}
