+++ Modules/PrimerCore/Sources/PrimerCore/Analytics/Models/PrimerPaymentHandling.swift
@@ -1,9 +1,10 @@
 //
-//  AnalyticsEventProtocol.swift
+//  PrimerPaymentHandling.swift
 //
 //  Copyright © 2026 Primer API Ltd. All rights reserved. 
 //  Licensed under the MIT License. See LICENSE file in the project root for full license information.
 
-@_spi(PrimerInternal) import PrimerCore
-
-extension Analytics.Event: AnalyticsEvent {}
+public enum PrimerPaymentHandling: String, Codable {
+    case auto   = "AUTO"
+    case manual = "MANUAL"
+}
