+++ Modules/PrimerCore/Sources/PrimerCore/Analytics/Models/PrimerSDKIntegrationType.swift
@@ -1,12 +1,12 @@
 //
 //  PrimerSDKIntegrationType.swift
 //
-//  Copyright © 2025 Primer API Ltd. All rights reserved. 
+//  Copyright © 2026 Primer API Ltd. All rights reserved. 
 //  Licensed under the MIT License. See LICENSE file in the project root for full license information.
 
 import Foundation
 
-internal enum PrimerSDKIntegrationType: String, Codable {
+@_spi(PrimerInternal) public enum PrimerSDKIntegrationType: String, Codable {
     case dropIn     = "DROP_IN"
     case headless   = "HEADLESS"
 }
