+++ Modules/PrimerCore/Sources/PrimerCore/Classes/Analytics.swift
@@ -1,9 +1,7 @@
 //
-//  AnalyticsEventProtocol.swift
+//  Analytics.swift
 //
 //  Copyright © 2026 Primer API Ltd. All rights reserved. 
 //  Licensed under the MIT License. See LICENSE file in the project root for full license information.
 
-@_spi(PrimerInternal) import PrimerCore
-
-extension Analytics.Event: AnalyticsEvent {}
+@_spi(PrimerInternal) public final class Analytics {}
