+++ Modules/PrimerCore/Sources/PrimerCore/Settings/DismissalMechanism.swift
@@ -1,9 +1,9 @@
 //
-//  AnalyticsEventProtocol.swift
+//  DismissalMechanism.swift
 //
 //  Copyright © 2026 Primer API Ltd. All rights reserved. 
 //  Licensed under the MIT License. See LICENSE file in the project root for full license information.
 
-@_spi(PrimerInternal) import PrimerCore
-
-extension Analytics.Event: AnalyticsEvent {}
+public enum DismissalMechanism: Codable {
+    case gestures, closeButton
+}
