+++ Modules/PrimerCore/Sources/PrimerCore/Settings/PrimerAppearanceMode.swift
@@ -1,9 +1,11 @@
 //
-//  AnalyticsEventProtocol.swift
+//  PrimerAppearanceMode.swift
 //
 //  Copyright © 2026 Primer API Ltd. All rights reserved. 
 //  Licensed under the MIT License. See LICENSE file in the project root for full license information.
 
-@_spi(PrimerInternal) import PrimerCore
-
-extension Analytics.Event: AnalyticsEvent {}
+public enum PrimerAppearanceMode: String, Codable {
+    case system = "SYSTEM"
+    case light = "LIGHT"
+    case dark = "DARK"
+}
