+++ Modules/PrimerCore/Sources/PrimerCore/Settings/PrimerDebugOptions.swift
@@ -1,0 +1,13 @@
+//
+//  PrimerDebugOptions.swift
+//
+//  Copyright © 2026 Primer API Ltd. All rights reserved. 
+//  Licensed under the MIT License. See LICENSE file in the project root for full license information.
+
+public struct PrimerDebugOptions: Codable {
+    public let is3DSSanityCheckEnabled: Bool
+
+    public init(is3DSSanityCheckEnabled: Bool? = nil) {
+        self.is3DSSanityCheckEnabled = is3DSSanityCheckEnabled != nil ? is3DSSanityCheckEnabled! : true
+    }
+}
