+++ Modules/PrimerCore/Sources/PrimerCore/Settings/PrimerKlarnaOptions.swift
@@ -1,0 +1,14 @@
+//
+//  PrimerKlarnaOptions.swift
+//
+//  Copyright © 2026 Primer API Ltd. All rights reserved. 
+//  Licensed under the MIT License. See LICENSE file in the project root for full license information.
+
+public final class PrimerKlarnaOptions: Codable {
+
+    public let recurringPaymentDescription: String
+
+    public init(recurringPaymentDescription: String) {
+        self.recurringPaymentDescription = recurringPaymentDescription
+    }
+}
