+++ Sources/PrimerSDK/Classes/Extensions & Utilities/Device.swift
@@ -1,0 +1,27 @@
+//
+//  Device.swift
+//
+//  Copyright © 2026 Primer API Ltd. All rights reserved. 
+//  Licensed under the MIT License. See LICENSE file in the project root for full license information.
+
+@_spi(PrimerInternal) import PrimerCore
+import UIKit
+
+extension Device {
+    static var deviceIdentifier: String {
+        // Prefer identifierForVendor
+        if let udid = UIDevice.current.identifierForVendor?.uuidString {
+            return udid
+        }
+        let userDefaults = UserDefaults.primerFramework
+        let udidKey = "Primer.uniqueDeviceIdentifier"
+        // If we have a previous UDID, use it
+        if let udid = userDefaults.string(forKey: udidKey) {
+            return udid
+        }
+        // If we have no previous UDID, create one
+        let udid = UUID().uuidString
+        userDefaults.setValue(udid, forKey: udidKey)
+        return udid
+    }
+}
