+++ Tests/Primer/Network/Services/NetworkingReportingServiceTests.swift
@@ -4,7 +4,7 @@
 //  Copyright © 2026 Primer API Ltd. All rights reserved. 
 //  Licensed under the MIT License. See LICENSE file in the project root for full license information.
 
-@_spi(PrimerInternal) @testable import PrimerSDK
+@testable import PrimerSDK
 @_spi(PrimerInternal) import PrimerCore
 import XCTest
 
@@ -101,7 +101,7 @@ final class NetworkingReportingServiceTests: XCTestCase {
             XCTAssertTrue(event?.properties is NetworkConnectivityEventProperties)
 
             let properties = event?.properties as! NetworkConnectivityEventProperties
-            XCTAssertEqual(properties.networkType, .wifi)
+            XCTAssertEqual(properties.networkType, Connectivity.NetworkType.wifi.rawValue)
             expectation.fulfill()
         }
